% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualMax.R
\name{annualMax}
\alias{annualMax}
\title{Annual maxima from time series}
\usage{
annualMax(dates, values, data, shift = 0, ...)
}
\arguments{
\item{dates}{Dates in ascending order.
Can be charater strings or \code{\link{strptime}} results,
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be mapped in color with \code{\link{colPoints}}}

\item{data}{Optional: data.frame with the column names as given by dates and values}

\item{shift}{Number of days to move the year-break to.
E.g. shift=61 for German hydrological year (Nov to Oct). DEFAULT: 0}

\item{\dots}{Further arguments passed to \code{\link{colPoints}}}
}
\value{
data.frame with numerical columns "year" and "max". 
The rownames are character strings with the year.
}
\description{
Get annual maxima from time series
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
annmax <- annualMax("date","Koeln", dis)
head(annmax)
plot(annmax, type="l", las=1, xlab="year")

head(annualMax("date","Koeln", dis))
head(annualMax("date","Koeln", dis, shift=117)) # hydrological year

}
\seealso{
\code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
