% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkShift.R
\name{checkShift}
\alias{checkShift}
\title{check shift values}
\usage{
checkShift(shift)
}
\arguments{
\item{shift}{Shifting value to be checked}
}
\value{
corrected shift value
}
\description{
internal function to check and correct shift values
}
\examples{
\dontrun{ # Excluded because it is internal
checkShift(-6)
checkShift(1:5)
checkShift(0)
checkShift(pi)
}


}
\seealso{
\code{\link{seasAxis}}, \code{\link{annualMax}}, \code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
