% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnames.R
\name{gnames}
\alias{gnames}
\title{gauge names}
\usage{
gnames(set = "all")
}
\arguments{
\item{set}{Character: Which subset of the gauge names should be returned? 
Can be abbreviated as per \code{\link{match.arg}}.
The options are described above. DEFAULT: "all"}
}
\value{
Vector with names (charstrings)
}
\description{
streamflow measurement gauge names along the rhine and tributaries.
With the \code{set} argument, different subsets can be chosen.
\tabular{llll}{
all(default)  \tab all 196 names in alphabetical order                \cr
large         \tab 55 gauges at large streams                         \cr
app           \tab large minus two stations with few data - for the \code{\link{rfsApp}} \cr
trend         \tab 35 gauges for the local trend app                  \cr
rawr,rawt     \tab 10 or 5 gauges along the rhine and tributaries - for the local raw data app          \cr
rhine         \tab 12 gauges along the actual Rhine river             \cr
poster        \tab 9 regime-representative gauges                     \cr
paper         \tab 4 gauges selected to be in the scientific article  \cr
}
}
\examples{
gnames()

}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
