% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs-package.R
\docType{data}
\name{map}
\alias{map}
\title{Europe borders map}
\format{SpatialPolygonsDataFrame}
\usage{
map
}
\description{
map: A dataset with borders of European countries used in the 
             \code{\link{rfsApp}}
}
\examples{
str(map@data)
}
\keyword{datasets}
