% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs-package.R
\docType{data}
\name{meta}
\alias{meta}
\title{meta data about gauges}
\format{A data.frame with 196 rows for 28 variables}
\usage{
meta
}
\description{
meta: A dataset with meta data about the streamflow gauges
}
\keyword{datasets}
