% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap.R
\name{minimap}
\alias{minimap}
\alias{plotmap}
\title{DEM plotting function}
\usage{
minimap(name, x1 = 0.1, x2 = x1 + 0.2, y1 = 0.7, y2 = y1 + 0.27,
  spargs = NULL, ...)

plotmap(name, metadf = meta, pch = 3, lwd = 3, col = "red",
  allargs = NULL, expr = NULL, label = FALSE, ...)
}
\arguments{
\item{name}{Station name(s) to be plotted with red crosses}

\item{x1, x2, y1, y2}{Relative location of minimap}

\item{spargs}{List of arguments passed to \code{\link{smallPlot}}. DEFAULT: NULL}

\item{\dots}{Further arguments passed from \code{minimap} to 
\code{plotmap} and then to \code{\link{points}}}

\item{metadf}{Dataframe with metadata. DEFAULT: meta}

\item{pch, lwd, col}{Point properties}

\item{allargs}{List of arguments passed to all gauge locations first}

\item{expr}{Expression to be executed after points, see example. DEFAULT: NULL}

\item{label}{Logical (vector): Label gauges? DEFAULT: FALSE}
}
\value{
None
}
\description{
Plots DEM as (small inset) map with desired stations highlighted
}
\examples{
plotmap("Koeln") ; axis(1) ; axis(2)
plotmap(c("Rees","Mainz","Worms","Seifen"), label=TRUE, col=1:4)
plot(1:10)
minimap("Koeln")
minimap(0) # no stations
minimap("Cochem", allargs=list(pch=3,cex=0.5))

stats <- gnames("app")[c(1,5,8,12)]
minimap(stats, expr=text(meta[name,"lon"], meta[name,"lat"], name, col=2, adj=c(0,1)))

op <- par(mfrow=c(2,2))
plot(1:10)
plot(1:10)
plot(1:10)
minimap("Koeln")
minimap("Mainz", spargs=list(outer=TRUE))
par(op)

pdf("test.pdf", height=5)
hist(rnorm(100))
minimap(c("Koeln","Rheinfelden"), y1=0.1, y2=0.9, x2=0.9)
minimap() # all dots
dev.off()
#berryFunctions::openFile("test.pdf")
unlink("test.pdf")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{hplot}
