% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyCompute.R
\name{qdoyCompute}
\alias{qdoyCompute}
\title{Quantile estimation per DOY}
\usage{
qdoyCompute(dates, values, data, shift = 0, RPs = RPs, ...)
}
\arguments{
\item{dates}{Dates in ascending order.
Can be charater strings or \code{\link{strptime}} results,
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be analyzed}

\item{data}{Optional: data.frame with the column names as given by dates and values}

\item{shift}{Number of days prior to Jan 1 to move the year-break to.
E.g. shift=61 for German hydrological year (Nov to Oct). DEFAULT: 0}

\item{RPs}{Return periods for \code{extremeStat::\link[extremeStat]{distLextreme}}
DEFAULT: \code{\link{RPs}}}

\item{\dots}{Further arguments, currently ignored}
}
\value{
Matrix for each DOY and quantile
}
\description{
Quantiles per Day of Year (doy), basically forked from
\url{https://github.com/brry/berryFunctions/blob/master/R/seasonality.R}
}
\examples{
# to be abstracted from testing

}
\seealso{
\code{\link{qdoyPeriods}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{chron}
\keyword{ts}
