% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyPeriods.R
\name{qdoyPeriods}
\alias{qdoyPeriods}
\title{DOY quantiles for several periods}
\usage{
qdoyPeriods(name, disdf = get("dis"), sans1999 = FALSE, shift = 117,
  RPs = RPvals, time = c(1920, 1950, 1980, 2010), ...)
}
\arguments{
\item{name}{Character: Name of gauge to be analyzed, see \code{\link{gnames}}}

\item{disdf}{Dataframe with columns "date" and \code{name}. DEFAULT: dis}

\item{sans1999}{Logical: exclude the values from 1999? (Pentecost flood)
DEFAULT: FALSE}

\item{shift, RPs}{Arguments passed to \code{\link{qdoyCompute}}}

\item{time}{Period beginnings (and ending of last one).
DEFAULT: c(1920,1950,1980,2010)}

\item{\dots}{Further arguments passed to \code{\link{qdoyCompute}}}
}
\value{
3d array for each doy, quantile and period
}
\description{
Quantiles per DOY for several time periods.
Enables analysis of seasonality changes over time.
}
\examples{
\dontrun{
load(seasFolder("data/dismeta.Rdata"))
qdoy <- qdoyPeriods("Koeln") # 3x3 seconds
str(qdoy)
qdoy <- list(Koeln=qdoy, Mainz=qdoyPeriods("Mainz", progbar=FALSE))
qdoyVisPeriods("Koeln", qdoy)
qdoyVisPeriods("Mainz", qdoy)
qdoyVisPeriods("Mainz", qdoy, sd=3)

qdoyNA <- list(Oberriet_Blatten=qdoyPeriods("Oberriet_Blatten", progbar=FALSE))
qdoyVisPeriods("Oberriet_Blatten", qdoyNA)
}

}
\seealso{
\code{\link{qdoyCompute}}, \code{\link{qdoyVis}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{chron}
\keyword{ts}
