% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyVis.R
\name{qdoyVis}
\alias{qdoyVis}
\title{Visualize seasonality of discharge}
\usage{
qdoyVis(qd, dist = "gev", RPs = c(1.111, 2, 10, 50, 200), cols = RPcols,
  main = dist, ylim = NA, add = FALSE, shift = 0, labelindex = 200,
  ...)
}
\arguments{
\item{qd}{Object returned by \code{\link{qdoyCompute}}}

\item{dist}{Distribution function: gev or empirical. DEFAULT: "gev"}

\item{RPs}{Return periods to be plotted (must be present in \code{qd}).
DEFAULT: c(1.111,2,10,50,200)}

\item{cols}{Colors for each line. Must have "RP.x" as names (or it is tried to set them).
DEFAULT: \code{\link{RPcols}}}

\item{main}{Plot title. DEFAULT: dist}

\item{ylim}{Y axis limits. DEFAULT: NA (computed internally from qd)}

\item{add}{Logical: add to existing plot?}

\item{shift}{Shift of year break, see \code{\link{qdoyCompute}}. DEFAULT: 0}

\item{labelindex}{Index along X axis at which to label lines.
Can be vectorized for each RP. 0 to suppress. DEFAULT: 200}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
Nothing, only plots
}
\description{
Visualize seasonality of discharge via quantile per DOY.
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
dd <- selectDates(1990,2010, df=dis)[,c("date","Koeln")]
qdoy <- qdoyCompute("date", "Koeln", data=dd, shift=117) # 2 secs
str(qdoy)
qdoyVis(qdoy, shift=117)

qdoyVis(qdoy, main="Cologne 1990-2010", RPs=50, cols=4, ylim=c(2e3,10e3), shift=117, lab=0)
qdoyVis(qdoy, dist="empirical", RPs=50, cols=3, add=TRUE, lab=0)
legend("topright", c("empirical", "gev"), col=3:4, lwd=3)

}
\seealso{
\code{\link{qdoyCompute}}, \code{\link{qdoyVisChange}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2017
}
\keyword{aplot}
