% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyVisPeriods.R
\name{qdoyVisPeriods}
\alias{qdoyVisPeriods}
\title{Visualize seasonality changes}
\usage{
qdoyVisPeriods(name, seaslist, qdp = seaslist[[name]], dist = "gev",
  metadf = get("meta"), steps = 1:3, RPs = c(1.111, 2, 10, 100),
  sd = -1, map = TRUE, axis = TRUE, legend = TRUE, x1 = 0.65,
  x2 = 0.95, y1 = 0.85, y2 = 1, text = TRUE, mar = c(1.8, 3.2, 3,
  0.2), cex.axis = 1, boxcol = 1, boxlwd = 1, maincex = 1,
  mainline = NA, mainadj = 0.35, mainsep = "\\n", xticks.lwd = 1,
  x.line = 0, ...)
}
\arguments{
\item{name}{Gauge station name}

\item{seaslist}{Named list as e.g. returned by \code{lapply(SOMENAMES, qdoyPeriods)}}

\item{qdp}{Array with qualtiles per doy and period as retuned by
\code{\link{qdoyPeriods}}. DEFAULT: \code{seaslist[[name]]}}

\item{dist}{Distribution function: gev or empirical. DEFAULT: "gev"}

\item{metadf}{Dataframe with metadata. DEFAULT: meta}

\item{steps, RPs}{Which of the periods and quantiles should be plotted?
DEFAULT: 1:3 (all of them), c(1.111,2,10,50,200)}

\item{sd}{Smoothing degree, see \code{\link{smoothFFT}}.
DEFAULT: negative value to suppress smoothing}

\item{map}{Should \code{\link{minimap}} be added? DEFAULT: TRUE}

\item{axis}{Should x axis with months be added? DEFAULT: TRUE}

\item{legend}{Should Period legend be added in topright? DEFAULT: TRUE}

\item{x1, x2, y1, y2}{Location of legend. See \code{\link{smallPlot}}}

\item{text}{Should lines be labeled with quantiles? DEFAULT: TRUE}

\item{mar}{Margins passed to \link{par}. DEFAULT: c(1.8,3.2,3,0.2)}

\item{cex.axis}{Size of axis labels (also y axis) and text. DEFAULT: 1}

\item{boxcol, boxlwd}{Color and width of box around plot region. DEFAULTS: 1,1}

\item{maincex, mainline, mainadj, mainsep}{Arguments to customize the title.}

\item{xticks.lwd, x.line}{X axis tick width and line. DEFAULT. 1,0}

\item{\dots}{Further arguments passed to \code{\link{qdoyCompute}}}
}
\value{
Nothing?
}
\description{
Visualize seasonality changes: compare periods
}
\examples{
# see qdoyPeriods

}
\seealso{
\code{\link{qdoyPeriods}}, \code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{hplot}
