% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfsApp.R
\name{rfsApp}
\alias{rfsApp}
\title{Launch interactive data analysis}
\usage{
rfsApp(app = "rhine", ...)
}
\arguments{
\item{app}{Folder specifying which app to run. 
DEFAULT: "rhine" (the main seasonality change app)}

\item{\dots}{Arguments passed to \code{shiny::\link[shiny]{runApp}}}
}
\description{
Launch interactive analysis of flood seasonality changes 
             along the Rhine river and tributaries.
             The R session is blocked during usage, close the app to re-enable
             console usage.
}
\examples{
# rfsApp()
}
\seealso{
\code{shiny::\link[shiny]{runApp}}, \code{\link{qdoyVisPeriods}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
\keyword{iplot}
