% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPs.R
\docType{data}
\name{RPs}
\alias{RPs}
\alias{RPcols}
\title{Return Periods}
\format{An object of class \code{numeric} of length 10.}
\usage{
RPs

RPcols
}
\value{
Numerical (RPs) or character string (RPcols) vector
}
\description{
Common Return period values
}
\examples{
RPs
plot(RPs, seq_along(RPs), log="x", col=RPcols, pch=16, cex=3)
round((1-1/RPs)*100,2)

}
\seealso{
\code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2017
}
\keyword{data}
\keyword{datasets}
