% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPvals.R
\docType{data}
\name{RPvals}
\alias{RPvals}
\alias{RPcols}
\title{Return Periods}
\format{An object of class \code{numeric} of length 10.}
\usage{
RPvals

RPcols
}
\value{
Numerical (RPvals) or character string (RPcols) vector
}
\description{
Common Return period values
}
\examples{
RPvals
plot(RPvals, seq_along(RPvals), log="x", col=RPcols, pch=16, cex=3)
round((1-1/RPvals)*100,2)

}
\seealso{
\code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2017
}
\keyword{data}
\keyword{datasets}
