% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs-package.R
\docType{data}
\name{seas}
\alias{seas}
\title{seas - streamflow aggregates}
\format{A list with 55 elements for 55 gauges (along large rivers, see \code{\link{gnames}}).
Each element has an array with 2 distributions, 11 return periods, 366 days and 3 periods.\cr
Included as a dataset in the package with code in R/rfs-package.R}
\usage{
seas
}
\description{
seas: A dataset with streamflow aggregates for seasonality analysis
}
\examples{
str(seas[[1]])
}
\keyword{datasets}
