% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasTrend.R
\name{seasTrend}
\alias{seasTrend}
\title{Seasonality trend 1960:2010}
\usage{
seasTrend(n, disdf = get("dis"), RP = 1.2, plot = TRUE, map = FALSE,
  trex = TRUE, peak = TRUE, shift = 61, ...)
}
\arguments{
\item{n}{Character: Name of gauge to be analyzed, see \code{\link{gnames}}}

\item{disdf}{Dataframe with columns "date" and \code{n}. DEFAULT: dis}

\item{RP}{Return Period for threshold. DEFAULT: 1.2}

\item{plot}{Logical: plot result? DEFAULT: TRUE}

\item{map}{Logical: draw \code{\link{minimap}}? DEFAULT: FALSE}

\item{trex, peak}{Logical: Plot trendline for (annual peaks of) threshold excesses}

\item{shift}{Shift passed to \code{\link{seasonality}}. DEFAULT: 61}

\item{\dots}{Further arguments passed to \code{\link{linReg}}}
}
\value{
Vector with RP, threshold, lm coefficients both for 
        all threshold exceedances and annual peaks of those
}
\description{
Seasonality trend 1960:2010
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
seasTrend("Koeln")

}
\seealso{
\code{\link{rfsApp}("trend")}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2017
}
\keyword{aplot}
