% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDates.R
\name{selectDates}
\alias{selectDates}
\alias{aliasMayBeRemoved}
\title{title}
\usage{
selectDates(y1 = 1920, y2 = 2010, m1 = 1, d1 = 1, m2 = 1, d2 = 1,
  df = get("dis"), logic = FALSE)
}
\arguments{
\item{y1, m1, d1}{Beginning year, month and day. DEFAULTS: 1920,1,1}

\item{y2, m2, d2}{Ending year, month and day. DEFAULTS: 2010,1,1}

\item{df}{Dataframe with discharge records in columns. DEFAULT: dis}

\item{logic}{Logical: return T/F vector: which rows of df are between a and b?}
}
\value{
ReturnValue
}
\description{
description
}
\details{
detailsMayBeRemoved
}
\section{Warning}{
 warningMayBeRemoved
}

\examples{
load(seasFolder("data/dismeta.Rdata"))
str(selectDates(y1=2009)[,c(1,14:18)])

}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{aplot}
