% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothFFT.R
\docType{data}
\name{smoothFFT}
\alias{smoothFFT}
\alias{smoothFFTexampleData}
\title{smooth time series with FFT}
\format{An object of class \code{numeric} of length 366.}
\usage{
smoothFFT(x, sd, ...)

smoothFFTexampleData
}
\arguments{
\item{x}{Numerical vector}

\item{sd}{Standard deviation. Determines degree of smoothing.
Use negative value to return original x.}

\item{\dots}{Further arguments passed to \code{fftw::\link[fftw]{IFFT}}}
}
\value{
Smoothed vector
}
\description{
smooth time series with FFT
}
\examples{
x <- cumsum(rnorm(200))
plot(x, type="l")
lines(smoothFFT(x, sd=1), col=4)
lines(smoothFFT(x, sd=4), col=2)

x <- smoothFFTexampleData
plot(x, type="l")
lines(smoothFFT(x, sd=4), col=4)
lines(smooth.spline(x,spar=0.45), col=2)

x <- x[1:100]
plot(x, type="l")
lines(smoothFFT(x, sd=3), col=4)    # works better for seasonal data, I guess
lines(smooth.spline(x,spar=0.45), col=2)

stopifnot(all(x == smoothFFT(x, -1)))

}
\seealso{
\code{\link{qdoyVis}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
\keyword{datasets}
\keyword{nonparametric}
\keyword{smooth}
\keyword{ts}
