% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STplot.R
\name{STplot}
\alias{STplot}
\alias{aliasMayBeRemoved}
\title{title}
\usage{
STplot(name, STlist = get("ST"), ylim = NA, map = TRUE)
}
\arguments{
\item{name}{Gauge name}

\item{map}{Logical. Should map be plotted? DEFAULT: TRUE}
}
\value{
ReturnValue
}
\description{
description
}
\details{
detailsMayBeRemoved
}
\section{Warning}{
 warningMayBeRemoved
}

\examples{
load(seasFolder("data/ST.Rdata"))
STplot("Koeln")

}
\seealso{
\code{\link{help}}, \code{graphics::\link[graphics]{plot}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2017
}
\keyword{aplot}
