% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs-package.R
\docType{data}
\name{thresfuns}
\alias{thresfuns}
\title{threshold computation functions}
\format{A list of functions}
\usage{
thresfuns
}
\description{
thresfuns: streamflow depending on return period 
             for 35 gauges along the Rhine. Used as threshold in \code{\link{seasTrend}}
}
\examples{
thresfuns$Koeln(5) # streamflow with return period of 5 years
}
\keyword{datasets}
