% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualMax.R
\name{annualMax}
\alias{annualMax}
\title{Annual maxima from time series}
\usage{
annualMax(dates, values, data, shift = 0, ...)
}
\arguments{
\item{dates}{Dates in ascending order.
Can be charater strings or \code{\link{strptime}} results,
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be mapped in color with \code{\link{colPoints}}}

\item{data}{Optional: data.frame with the column names as given by dates and values}

\item{shift}{Number of days to move the year-break to.
E.g. shift=61 for German hydrological year (Nov to Oct). DEFAULT: 0}

\item{\dots}{Further arguments passed to \code{\link{colPoints}}}
}
\value{
named numeric vector
}
\description{
Get annual maxima from time series
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
annual_max <- annualMax("date","Koeln", dis)
head(data.frame(annual_max))
plot(as.numeric(names(annual_max)), annual_max, type="l", las=1, xlab="year")

}
\seealso{
\code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
