% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualMax.R
\name{annualMax}
\alias{annualMax}
\title{Annual maxima from time series}
\usage{
annualMax(dates, values, data, shift = 0, missing2NA = TRUE)
}
\arguments{
\item{dates}{Dates in ascending order.
Can be charater strings or \code{\link{strptime}} results,
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be mapped in color with \code{\link{colPoints}}}

\item{data}{Optional: data.frame with the column names as given by dates and values}

\item{shift}{Number of days to move the year-break to.
E.g. shift=61 for German hydrological year (Nov to Oct). DEFAULT: 0}

\item{missing2NA}{Logical. If years are completely missing in \code{dates},
insert NA rows for those? DEFAULT: TRUE}
}
\value{
data.frame with one row per (hydrological) year (h_year) with the columns:\cr
year, max, date, doy (day of h_year), n (number of non-NA values), 
days (number of days per h_year in input data) and rownr.\cr
The rownames of the output are character strings with the h_year.
}
\description{
Get annual maxima and more information from time series
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
annmax <- annualMax("date","Koeln", dis)
head(annmax)
plot(annmax[,1:2], type="l", las=1, xlab="year")
plot(annmax$year, annmax$doy)

head(annualMax("date","Koeln", dis))
head(annualMax("date","Koeln", dis, shift=117)) # hydrological year
berryFunctions::seasonality("date","Koeln", dis, shift=117)
annmax <- annualMax("date","Koeln", dis, shift=117)
berryFunctions::linReg(doy~year, data=annmax); abline(h=150)
berryFunctions::linReg(doy~year, data=annmax[120:190,]); abline(h=150)
# No real trend in timing of annual streamflow maximum in Cologne

dis2 <- dis[dis$Koeln>9000,]
head(annualMax("date","Koeln", dis2))
head(annualMax("date","Koeln", dis2, missing2NA=FALSE))
#View(dis[,c("date","Koeln")])
plot(head(dis[,c("date","Koeln")], 500), type="l")

}
\seealso{
\code{\link{qdoyCompute}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
