% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnames.R
\name{gnames}
\alias{gnames}
\title{gauge names}
\usage{
gnames(large = FALSE, rhine = FALSE, poster = FALSE, paper = FALSE)
}
\arguments{
\item{large, rhine, poster, paper}{Logical: return only a subset of the gauge names? DEFAULT: FALSE}
}
\value{
Vector with names (charstrings)
}
\description{
streamflow measurement gauge names along the rhine and tributaries.
The default is to return all 196 names in alphabetical order.
With the optional arguments, subsets can be chosen:\cr
large: 55 gauges at large streams\cr
rhine: 12 gauges along the actual Rhine river\cr
poster: 9 regime-representative gauges\cr
paper: 4 gauges selescted to be in the scientific article\cr
}
\examples{
gnames()

}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
