% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap.R
\name{minimap}
\alias{minimap}
\title{DEM plotting function}
\usage{
minimap(name, metadf = meta, x1 = 0.1, x2 = x1 + 0.2, y1 = 0.7,
  y2 = y1 + 0.27, pch = 3, lwd = 3, col = "red", allargs = NULL, ...)
}
\arguments{
\item{name}{Station name(s) to be plotted with red crosses}

\item{metadf}{Dataframe with metadata. DEFAULT: meta}

\item{x1, x2, y1, y2}{Relative location of minimap}

\item{pch, lwd, col}{Point properties}

\item{allargs}{List of arguments passed to all gauge locations first}

\item{\dots}{Further arguments passed to \code{\link{points}}}
}
\value{
None
}
\description{
Plots DEM as small inset map with desired stations highlighted
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
plot(1:10)
minimap("Koeln")
minimap(0) # no stations
minimap("Cochem", allargs=list(pch=3,cex=0.5))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{hplot}
