% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyCompute.R
\name{qdoyCompute}
\alias{qdoyCompute}
\title{Quantile estimation per DOY}
\usage{
qdoyCompute(dates, values, data, shift = 0, probs = c(0, 25, 50, 75, 95,
  99.9)/100, width = 3, quiet = FALSE, ...)
}
\arguments{
\item{dates}{Dates in ascending order.
Can be charater strings or \code{\link{strptime}} results,
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be mapped in color with \code{\link{colPoints}}}

\item{data}{Optional: data.frame with the column names as given by dates and values}

\item{shift}{Number of days to move the year-break to.
E.g. shift=61 for German hydrological year (Nov to Oct). DEFAULT: 0}

\item{probs}{Probabilities passed to \code{\link{quantileMean}} for plot=4.
DEFAULT: c(0,25,50,75,95,99)/100}

\item{width}{Numeric: window width for plot=4. Used as sd in gaussian weighting.
Support (number of values around a DOY passed to
quantile funtion at least once) is ca 4.9*width.
The value at doy itself is used 10 times.
Larger values of width require more computing time.
DEFAULT: 3}

\item{quiet}{Logical: suppress progress bar? DEFAULT: FALSE}

\item{\dots}{Further arguments, currently ignored}
}
\value{
Matrix for each DOY and quantile
}
\description{
Quantiles per Day of Year (doy), basically forked from
\url{https://github.com/brry/berryFunctions/blob/master/R/seasonality.R}
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
qdoy <- qdoyCompute("date", "Koeln", data=dis)
str(qdoy)
plot(1:366, qdoy[,6], ylim=c(400,11000), type="l")
lines(1:366, qdoy[,5], col=2)

}
\seealso{
\code{\link{qdoyPeriods}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{chron}
\keyword{ts}
