% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdoyPeriods.R
\name{qdoyPeriods}
\alias{qdoyPeriods}
\title{DOY quantiles for several periods}
\usage{
qdoyPeriods(name, disdf = get("dis"), sans1999 = FALSE, probs = c(0, 0.5,
  0.9, 0.99), width = 7, shift = 117, time = c(1920, 1950, 1980, 2010),
  ...)
}
\arguments{
\item{name}{Character: Name of gauge to be analyzed, see \code{\link{gnames}}}

\item{disdf}{Dataframe with columns "date" and \code{name}. DEFAULT: dis}

\item{sans1999}{Logical: exclude the values from 1999? (Pentecost flood)
DEFAULT: FALSE}

\item{probs, width, shift}{Arguments passed to \code{\link{qdoyCompute}}}

\item{time}{Period beginnings (and ending of last one).
DEFAULT: c(1920,1950,1980,2010)}

\item{\dots}{Further arguments passed to \code{\link{qdoyCompute}}}
}
\value{
3d array for each doy, quantile and period
}
\description{
Quantiles per DOY for several time periods.
Enables analysis of seasonality changes over time.
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
qdoy_Koeln <- qdoyPeriods("Koeln")
str(qdoy_Koeln)

}
\seealso{
\code{\link{qdoyCompute}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{chron}
\keyword{ts}
