% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs-package.R
\docType{package}
\name{rfs}
\alias{rfs}
\alias{rfs-package}
\alias{rfs-package}
\title{Rhine flood seasonality}
\description{
Rhine flood seasonality analysis.
PhD project by Berry Boessenkool.
ToDo: add more info here
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2017
}
\keyword{documentation}
\keyword{package}
