% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasFolder.R
\name{seasFolder}
\alias{seasFolder}
\title{Get path to folder with seasonality analysis + data}
\usage{
seasFolder(...)
}
\arguments{
\item{\dots}{Optional filename(s) that will be appended.}
}
\value{
Path(s)
}
\description{
Get path to (files in) the folder with seasonality analysis + data
}
\examples{
seasFolder()
seasFolder("dummy.txt")
stopifnot(length(seasFolder("dummy.txt", "dude.xyz"))==2)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{file}
