% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasFolder.R
\name{seasFolder}
\alias{seasFolder}
\title{Get path to folder with seasonality analysis + data}
\usage{
seasFolder(..., warnonly = FALSE, path = NA)
}
\arguments{
\item{\dots}{Optional filename(s) that will be appended.}

\item{warnonly}{Logical: only warn instead of stopping if files are not found. 
NA to not check file existence. DEFAULT: FALSE}

\item{path}{Character: path to search in. DEFAULT: NA}
}
\value{
Path(s)
}
\description{
Get path to (files in) the folder with seasonality analysis + data
}
\examples{
seasFolder()
seasFolder("dummy.txt", warnonly=TRUE)
stopifnot(length(seasFolder("dummy.txt", "dude.xyz", warnonly=TRUE))==2)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{file}
