% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasTrend.R
\name{seasTrend}
\alias{seasTrend}
\title{Seasonality trend 1960:2010}
\usage{
seasTrend(n, disdf = get("dis"), RP = 1.2, plot = TRUE, map = FALSE,
  trex = TRUE, peak = TRUE, startmonth = 1, nmonths = 12,
  legpos = "left", shift = 61, seasargs = NULL, ...)
}
\arguments{
\item{n}{Character: Name of gauge to be analyzed, see \code{\link{gnames}}}

\item{disdf}{Dataframe with columns "date" and \code{n}. DEFAULT: dis}

\item{RP}{Return Period for threshold. DEFAULT: 1.2}

\item{plot}{Logical: plot result? DEFAULT: TRUE}

\item{map}{Logical: draw \code{\link{minimap}}? DEFAULT: FALSE}

\item{trex}{Logical: Plot trend for threshold excesses? DEFAULT: TRUE}

\item{peak}{Logical: Plot trend for annual peaks of threshold excesses? DEFAULT: TRUE}

\item{startmonth}{Integer between 1 and 12. To select which months
are used in the trend analysis. DEFAULT: 1}

\item{nmonths}{Integer between 1 and 12. Hom many months are used? DEFAULT: 12}

\item{legpos}{Position of \code{\link{legend}}. DEFAULT: "left"}

\item{shift}{Shift passed to \code{\link{seasonality}}. DEFAULT: 61}

\item{seasargs}{List of arguments passed to \code{\link{seasonality}}, 
like e.g. seasargs=list(xaxt="n"). DEFAULT: NULL}

\item{\dots}{Further arguments passed to \code{\link{linReg}}}
}
\value{
Vector with RP, threshold, lm coefficients both for 
        all threshold exceedances and annual peaks of those
}
\description{
Seasonality trend 1960:2010
}
\examples{
load(seasFolder("data/dismeta.Rdata"))
seasTrend("Koeln")
seasTrend("Koeln", seasargs=list(xaxt="n", tlab="Empty"))

}
\seealso{
\code{\link{rfsApp}("trend")}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2017
}
\keyword{aplot}
