% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothFFT.R
\docType{data}
\name{smoothFFT}
\alias{smoothFFT}
\alias{smoothFFTexampleData}
\title{smooth time series with FFT}
\format{An object of class \code{numeric} of length 366.}
\usage{
smoothFFT(x, sd, keepNA = TRUE, quiet = FALSE, ...)

smoothFFTexampleData
}
\arguments{
\item{x}{Numerical vector}

\item{sd}{Standard deviation. Determines degree of smoothing.
Use negative value to return original x.}

\item{keepNA}{Logical: keep NA values? For computation, NAs will be
interpolated with
\code{berryFunctions::\link[berryFunctions]{approx2}}.
If keepNA=TRUE, the result will again have NAs.
If keepNA=FALSE, they will remain interpolated. DEFAULT: TRUE}

\item{quiet}{Logical. Suppress warnings about all-NA vectors? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{fftw::\link[fftw]{IFFT}}}
}
\value{
Smoothed vector
}
\description{
smooth time series with FFT
}
\examples{
x <- cumsum(rnorm(200))
plot(x, type="l")
lines(smoothFFT(x, sd=1), col=4)
lines(smoothFFT(x, sd=4), col=2)

x[105:111] <- NA
plot(x, type="l")
lines(smoothFFT(x, sd=3), col=4, lwd=2)
lines(smoothFFT(x, sd=3, keepNA=FALSE), col=2, lty=3)

x <- smoothFFTexampleData
plot(x, type="l")
lines(smoothFFT(x, sd=4), col=4)
lines(smooth.spline(x,spar=0.45), col=2)

x <- x[1:100]
plot(x, type="l")
lines(smoothFFT(x, sd=3), col=4)    # works better for seasonal data, I guess
lines(smooth.spline(x,spar=0.45), col=2)

smoothFFT(c(42,NA,NA,NA,NA)[-1], 2)

stopifnot(all(x == smoothFFT(x, -1)))

}
\seealso{
\code{\link{qdoyVis}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
\keyword{datasets}
\keyword{nonparametric}
\keyword{smooth}
\keyword{ts}
