% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STplot.R
\name{STplot}
\alias{STplot}
\title{Plot seasonality trend depending on threshold for "flood"}
\usage{
STplot(name, STlist = get("ST"), ylim = NA, map = TRUE)
}
\arguments{
\item{name}{Character string: Gauge name}

\item{STlist}{List as created in \code{\link{rfs-package}.R}. DEFAULT: ST}

\item{ylim}{Two numbers: Y axis limits. DEFAULT: NA (computed internally)}

\item{map}{Logical. Should map be plotted? DEFAULT: TRUE}
}
\description{
Plot seasonality trend depending on threshold for "flood"
}
\examples{
load(seasFolder("data/ST.Rdata"))
STplot("Koeln")

}
\seealso{
\code{\link{seasTrend}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2017
}
\keyword{aplot}
\keyword{hplot}
